requirements.txt
aiogram==3.0.0b7
openai>=1.12.0
pydantic==1.10.4
python-dotenv
pymongo
